fix(otel): align expectations to v0.150 upstream config defaults
- debug exporter: build expected from CreateDefaultConfig() + translator overrides (debugexporter QueueConfig has non-zero defaults a hardcoded struct missed)
- filterprocessor: compare exported config fields individually; Config now holds unexported ottl.Factory pointer maps that differ by address per CreateDefaultConfig()
- journald exporter: LogRetention is int32 under v0.150 cwlogs (SDK v2); cast retention and update the test expectation

+++ translator/translate/otel/exporter/journald/translator.go
@@ -110,7 +110,7 @@ func (t *translator) setJournaldFieldsFromConfig(cfg *awscloudwatchlogsexporter.
 	}
 
 	if retentionInDays, ok := t.collectConfig["retention_in_days"].(float64); ok {
-		cfg.LogRetention = int64(retentionInDays)
+		cfg.LogRetention = int32(retentionInDays)
 	}
 
 	return nil
